requirements.txt
Django==5.2.5
djangorestframework==3.16.1
Pillow==10.4.0
django-filter==24.1
